Skip one functional test
Skip this test until Bug #1745003 is fixed.  It's not a normal
skip; the skipping is done by commenting out the offending
assertions.  That's because I725c228b82efd408b283854c1be2d1bdab33a834
appears to be completely ineffective, probably because of the
mixins used in the db functional tests.

Change-Id: I01bc954fc524d64cfeae67f8d03fe15f998893e6
Related-bug: #1745003