Use configured value for import-methods header
In Image API v2.6, the image-create response returns an
'OpenStack-image-import-methods' header containing a list of
methods available in that cloud.  This patch populates that
list from the enabled_import_methods configuration setting.
Following RFC 7230, if the enabled_import_methods is empty,
the header is not returned with the response.

Change-Id: I3066da36a0fbb83fe492bddb193975e42ffc5084
Closes-bug: #1748559
(cherry picked from commit 3712dccfdb7ee30243e4efc09fcad1c3197bdb51)