Make the Image status transition early
When image-import call is made, all kind of things happens before
the image status transitions to 'importing'. We should do this as early
as possible to ensure that there is no user action against the image
while it's prcessed and the user gets the info that the importing is
actually happening.

Closes-Bug: #1749525

Change-Id: I3ff9f7af2f266dd60fd3ad738b1fb9eeb6fd9fb8