Hide Pike note on Queens Release Notes page
The pike-rc-2 release note was merged to master after the
stable/pike branch was cut and is appearing on the Queens
Release Notes page.  Take advantage of the reno ignore-notes
feature to hide it in Queens.  (It is still displayed on the
Pike Release Notes page, as it should be.)

Change-Id: Ib56104ecf3f94eec56e144a9ea6569fa9bf14b8e