Make eventlet monkey patching conform to best practices
Per [1], eventlet monkey patching should happen as early as
possible to avoid mismatches where a module was imported both before
and after it was monkey patched.  This is an exception to the import
order rules that should maybe be more explicitly called out.

In addition, partial monkey patching can be a problem, as shown in
the discussion of the thread module from the same document.  This
seems to be contributing to the doc build problems that are
occurring with the latest version of openstackdocstheme, so the
partial monkey patching is also removed in favor of full patching.

Change-Id: I0d2d9fb9f0b9d747ad1d955420f6ad129ebbfbcf
1: http://specs.openstack.org/openstack/openstack-specs/specs/eventlet-best-practices.html#monkey-patching
Closes-Bug: 1759935