Add time-bounded wait-for-status to func tests
Replace strategy of time.sleep(s_value) waiting for an image to
reach 'active' with a time-bounded wait-for-status loop function.
That way we can have a more generous s_value, which seems to
sometimes be needed in the gates, without having to use that much
time on every run.

I picked the initial s_value as 4x what was in the current code,
though given that some of the other gate tests run >1 hour, we could
easily make it 600 sec without lengthening the current minimum
time-to-verification of a patch.

Closes-bug: #1767142
Change-Id: Ib771079348d22212c7b6d90fa89e7fd4161422d3