Add 'useful image properties' document
Move the list of useful image properties from the glanceclient
docs to the Glance Admin Guide.  It makes more sense in the Glance
docs as the only connection they have with the glanceclient is that
you can use the client CLI to set these properties

The table of property keys/values is unchanged; the introductory
section is new.

Change-Id: I0c307a067490c06728b5adf70c91586254e337da
Needed-by: https://review.openstack.org/565782