Refactor wait_for_scrubber_shutdown function
The wait_for_scrubber_shutdown function catches AssertionErrors
and retries.  Refactor it so that it only retries for a specific
error (namely, when glance scrubber is running in daemon mode when
the test is being executed) by changing the function so that it
returns values instead of relying on test assertions in the
function passed to it.  Also refactor the tests that call
wait_for_scrubber_shutdown so that they check the returned results
instead of using assertions in the function they pass to
wait_for_scrubber_shutdown.

This is a follow-up to https://review.openstack.org/#/c/566681/

Change-Id: I7108179e0d96e09638ff783b029a8216f0938c3b
Closes-bug: #1768077