Add periodic tips jobs
Add jobs to the periodic queue that run the glance unit and
functional tests against the master of various libraries we
consume so that we don't have to wait for a release to detect
a change that could be problematic.

The dependency is needed for the *glance_store-tips jobs to
be processed properly.

Change-Id: I0385646991140debc06ea27a194a6e7f6f749e6b
Depends-On: https://review.openstack.org/#/c/571005/