Multiple backend support for scrubber
In Rocky multiple backend support is added as an experimental API
feature. As of now multiple backend works with glance-api but it
is missing in glance-scrubber.

Made provision to initialize multiple backend related config options
in scrubber.

NOTE:
Operator needs to define multiple backends related config options in
glance-scrubber.conf.

Change-Id: I7327ff5db8fa19e6813a0fe0fb657eab19f8b61f
Closes-Bug: #1794225