Clarify the Glance/metadefs relationship
I've answered a few questions recently for people who assumed that
since the metadefs service lives on the Glance endpoint, metadefs
have some kind of special relationship to images.  I think the
metadefs section of the api-ref makes it clear that this is not the
case (though it's not mentioned at all on the Images part of the
api-ref).  So here's an attempt to clarify issue (without being too
intrusive) right at the top of the Glance docs landing page.

Change-Id: I3dfbb09862acedb265a5a4557749bf52f6add5f4