Failure in web-dowload kept image in importing state
As of now if import using web-download method fails due to invalid
uri, image remains in 'importing' state. As the image is in 'importing'
state it is not of any use to end-user.

Resetting it to queued state, so that user can try again uploading/importing
data to the same image.

Change-Id: Iaa26922a9cf5a419c3da23bdfc347a6fdf6cb549
Closes-Bug: #1803299