Guard __getattr__ on QuotaImageTagsProxy
In the related bug (below) we can see getting tags from a
QuotaImageTagsProxy can sometimes cause an infinite loop, presumably
because of bizarre threading circumstances. To at least provide a bit
more information when this bizarre circumstance happens, instead
of looping, resulting in a stackoverflow and the processing being
terminated without a traceback, now raise a TypeError. This should
help figure out what may be going on in the rare times this happens.

Also, it is generally a healthy idea to always have a guard like this
in a __getattr__ that uses an attr.

Change-Id: Ib4ad119a838d5b06372d9302dc6608da1772eed6
Related-Bug: #1808063