Update show_multiple_locations deprecation note
Update logged text to indicate our intentions with respect
to the show_multiple_locations configuration option, which
has been deprecated since Newton and advertised as subject
to removal in Pike.  Also correct the misleading impression
in the deprecation notice and earlier release notes that
functionality that requires show_multiple_locations = True
can be configured via policies at the current time.  Also
adds a release note to this effect.

Not a straight cherry-pick.  Changed the section of the
relase note from 'upgrade' to 'issues' because this is more
of a "known issue" in the stable branches.

Change-Id: I5bf0c8af9dfe87e0d17f7a16d4676f387b4379f6
Closes-bug: #1808375
(cherry picked from commit de72ac4a01be1ccc75825eb941dad64c170d6bea)