Unit tests fails with new glance_store version 0.29.0
Around 11 unit tests are failing [1] with a recent patch [2] merged in
glance_store. The reason behind the failure is glance unit tests are
loading rbd store which is not supported in gate. Earlier these tests
were passing because they were not trying to connect to rados while
loading the store, but after merging of recent patch [1] it tries to
connect to rados via configure_add method of rbd driver.

To avoid these kind of failures in near future we should drop using
rbd store in tests as we know it is not supported in gate.

NOTE:
Also corrected the naming conventions of store identifiers, instead
of using file1 etc., used appropriate names.

[1] http://logs.openstack.org/40/663740/1/check/cross-glance-py27/50a9a9c/
[2] https://review.opendev.org/660340

Change-Id: I18add352238e93c5d4ade8b3e8d967b2f639cdad
Closes-Bug: #1831963