MultiStore: Avoid potential KeyError in functional tests
For glance multiple store, image shows backend/store information
in location metadata. It is accessed as a dictionary attribute, which can
lead to KeyError if 'backend' key is not present in location metadata.

To avoid this used get() call of dictionary which returns None if key
is not present in the dictionary.

Change-Id: I0f7fe80f13e79137c7bdaf5ec93e775c88b70cce
Closes-Bug: #1832490