MultiStore: Lazy update fails if image is not owned by owner
Lazy updating store information in location metadata fails if image
is not owned by the user. The reason is saving this information to
image object returns Forbidden error as the user is not allowed to change
the image information. Actually after updating the location information
there is no need to explicitly call save function because when location
information is updated using image.locations object it is updated to
database at the same time.

To fix this issue, remved the image_repo.save call, so that Lazy update
should be allowed to index as well as show call for all the users.

Change-Id: I92c85647ea4eea7069f8124334402f9127bf653c
Closes-Bug: #1840885