Correct the deprecation messages of local dir config
The deprecation messages of node_staging_uri and work_dir
were claiming that local directory is not needed after moving
to multistore. This is not exactly true and gave false
expectations of the current capabilities.

For now only the way to configure those needed folder and
how they are internally accessed is changing.

This change corrects the deprecation message to reflect
current state.

Change-Id: I39c170903c64181841a22c7b9bdaa3b5a1346caa
Closes-Bug: #1843891