Configure Glance TCP socket after the store initialization
When multiple stores are being used with rbd backend, Glance tries
to get fsid of the cluster using rados library.
If even one of the rbd backends is unavailable, glance-api wsgi fails
to start properly, but continue listening on the TCP socket and is not
responding at any request until the timeout.

This change moves socket configuration after the store initialization
to make sure that failed Glance API process will not be
considered as up by the haproxy TCP check.

Change-Id: I4298bb835ec60b79a0010e52318ada19eb24ad80
Related-Bug: #1861723