Remove configs and entries for deprecated registry
Including removal of the example conf in docs.
Including glance-registry command entrypoint to eventlet.
Including rpc_controller from wsgi, changed to reject.
Not including the files devstack is depending on.

This change means that glance-registry starts and announces
that it has been removed.
This change means that when ran as wsgi app, all calls to
/rpc endpoint will be rejected.

This will allow devstack to make a graceful transition to
life without the glance registry, which was deprecated in
Newton release.

Change-Id: I7bf3284cba4c38605fb50b3c458e53f896f34086