Remove usages of SSL related config options in utils
SSL support was removed by [1] which resulted in removal of some
SSL related config options, but two config options 'key_file' and
'cert_file' are still used in utils.stash_conf_values() which is called
in reload functionality. When SIGHUP is sent to the parent process,
this usage causes it to crash and its children processes become orphaned,
which then causes the reload test to be stale (fail).

This change also includes:
- Remove 2 attributes and the decorator that help to skip this test
  (self.diable, self.disabled_message, and @skip_if_disabled)

[1] https://review.opendev.org/#/c/697969/5/glance/common/wsgi.py

Change-Id: I361c9d0b53a67b1cc60f0f3012e22accdcc46055
Closes-Bug: #1855708