Install all deps in venv creation
The venv target was inheriting the default deps, which appropriately
just includes the main project requirements and our test requirements.
We also use venv for reno creation and other miscellaneous things, so it
is good to get all potential requirements installed. This adds the
installation of our docs requirements in the creation of the venv
environment.

Change-Id: I8b232116e54e39243a009a195090f12ef1a56bd4
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>