Make import task capable of running as admin on behalf of user
This makes the api_image_import task capable of running as an admin on
behalf of a user if so authorized by the API. It includes a new object
called ImportActionWrapper which provides a bundle of utility methods
which can be run either against the user-authorized or admin-authorized
ImageRepo passed in from the API. It encapsulates all the actions
we are able and willing to run as an admin for the user.

This is currently not drivable by the API because the policy check is
still statically defined as "admin or owner" but this change is
offered without any needed modification to the functional tests to
prove that it does not regress existing functionality. The following
patch will introduce a more robust knob for allowing users to do this,
and it brings the functional test changes with it.

Changes:
 - Removed the change to the images API to pass in an admin repo
 - Reinstated the ImportActionWrapper usage that was removed from
   backport commit 2b3ce07c32624f9987df87520c5ff8d538c8e5fd

Change-Id: Iac75956e314ec6f41db18430486bd8be9754e780
(cherry picked from commit 2fd0c2573311db6a50228ed22adf8bc2b5638cfe)