Add a test to replicate the owner-required behavior of copy-image
Right now, image_import(method='copy-image') requires you to be the owner
of the image. This adds a test that validates that behavior so we can measure
the difference of the next set of changes to modify that.

Change-Id: I284271a458117b753a88b847598ca10ce4fb94fa
(cherry picked from commit f450b9cf0314afeb9aa64ed687218b92b6a49321)