Add a policy knob for allowing non-owned image copying
This adds a copy_image policy knob which can be used to grant users
the ability to copy images other than their own using the new
functionality just added to the api_image_import task. By default,
only admins are allowed to do this.

A functional test modification is included to show that users can
be granted permission to do this based on something like the
"public visibility" attribute of an image.

Change-Id: Idebf66e2944bcddb7a5c76b81e47c654b401c2a8