Add context.elevated() helper for getting admin privileges
This adds an elevated() method to glance.context.RequestContext similar
to what Nova and other projects use. When doing something as admin on
behalf of a user, this results in a whole context, including information
about the user and the request, but with is_admin==True.

Change-Id: I5499946425b1c32476c57241b4b14b601daa841f
(cherry picked from commit c59ed1bce82c960b205663ba9ac6d038633588dc)