Add 'all' visibility filter for listing images in docs
Since this commit https://github.com/openstack/glance/commit/cfa9b6e8e814a52181d4e27fd23a659e8489dee8,
the visibility all is supported in the listing of images. However,
the documentation does not present it. This can make it hard for operators
to be aware of this option when listing images. Therefore, we propose to add
it in the documentation.

Change-Id: Iaa0ce6e1fde5c4c546b2f4a53969287b05d295de