Flesh out FakeImage for extra_properties
The FakeImage in test_images_resource.py is missing a lot of standard
properties. That becomes a problem as soon as you add extra_properties,
which gets delegated to for any missing property by ImageTarget. Thus,
if that fixture has extra_properties defined on it, any other base
property (i.e. checksum, size, etc) will be looked for inside that dict
if it is not defined on the fixture itself.

Thus, this adds fake values for all those properties and adds
extra_properties={} so that subsequent patches can put things into
extra_properties without a bunch of work.

Change-Id: I4b63706927b599eb343b614f0c396f699378e14b