Stream-friendly disk format inspection module
This adds a module called common.format_inspector which contains
an implementation of virtual disk format and size detection that
works in a streamable fashion. This can be injected into a data
processing loop and gracefully detect both format type as well as
virtual_size of the disk format, for those that are supported.

Currently excluded from the list of formats that glance claims to
support are iso (which can be treated like raw) and ploop (which
not even qemu-img seems to support).

This also adds a test_format_inspector.py executable into tools/,
which we can use to validate the code manually on images in the
case where bugs report that this fails to work on some real-world
image.

Related to blueprint calculate-virtual-size

Change-Id: Ica9fc645229b0b19b913ce3f2e151f7b0c935649